refactor(validator): move v1 schema from pkg/api/validator/v1 to pkg/validator/v1
The validator-catalog wire schema lived under pkg/api/ but pkg/api/ is
documented in docs/integrator/public-api.md as Internal HTTP-handler
code ("Server-side REST handlers (server-internal; consumers use the
HTTP API, not the Go types)"). The v1 schema isn't HTTP-specific — it's
the YAML wire format the validator package consumes — so its location
mis-labels its actual stability tier and forces sibling-into-internal
imports from pkg/validator/catalog → pkg/api/validator/v1.

Move pkg/api/validator/v1/ → pkg/validator/v1/ via git mv (history
preserved on all 14 files). Package name unchanged (still `v1`); only
import path changes. Symmetric with the pkg/client/v1 facade location
established by #1111 — both are now pkg/<domain>/v1/.

External Go importers reaching into pkg/api/validator/v1 (despite the
Internal label in public-api.md) need to update:

  - import v1 "github.com/NVIDIA/aicr/pkg/api/validator/v1"
  + import v1 "github.com/NVIDIA/aicr/pkg/validator/v1"

In-tree imports (13 files across pkg/cli, pkg/client/v1, pkg/validator,
validators/) updated mechanically.

No code changes; pure rename + import path update. Build, lint, and
tests clean on all affected packages.
