+++ pkg/client/v1/aicr.go
@@ -83,13 +83,13 @@ import (
 	"sync"
 	"time"
 
-	validatorv1 "github.com/NVIDIA/aicr/pkg/api/validator/v1"
 	"github.com/NVIDIA/aicr/pkg/constraints"
 	"github.com/NVIDIA/aicr/pkg/defaults"
 	"github.com/NVIDIA/aicr/pkg/errors"
 	"github.com/NVIDIA/aicr/pkg/recipe"
 	"github.com/NVIDIA/aicr/pkg/snapshotter"
 	"github.com/NVIDIA/aicr/pkg/validator"
+	validatorv1 "github.com/NVIDIA/aicr/pkg/validator/v1"
 	"sigs.k8s.io/yaml"
 )
