+++ pkg/validator/v1/affinity.go
