+++ pkg/validator/v1/affinity_test.go
