+++ pkg/validator/v1/catalog_test.go
