+++ validators/conformance/robust_controller_check_test.go
@@ -19,8 +19,8 @@ import (
 	"strings"
 	"testing"
 
-	v1 "github.com/NVIDIA/aicr/pkg/api/validator/v1"
 	"github.com/NVIDIA/aicr/pkg/recipe"
+	v1 "github.com/NVIDIA/aicr/pkg/validator/v1"
 	"github.com/NVIDIA/aicr/validators"
 	k8sfake "k8s.io/client-go/kubernetes/fake"
 )
